change int to float

+++ cuda/cuda_saxpy.cu
@@ -4,33 +4,33 @@
 #define NUM_BLOCKS 8  
 
 __global__
-void saxpy(int n, float a, int * x, int * y, int * out)
+void saxpy(int n, float a, float * x, float * y, float * out)
 {
   int i = blockIdx.x*blockDim.x + threadIdx.x;
   if (i < n) out[i] = a*x[i] + y[i];
 }
 
 
-void fill_array(int *data, int N){
+void fill_array(float * data, int N){
     for(int idx=0;idx<N;idx++){
         data[idx] = idx;
     }
 }
 
-void print_output(int *x, int *y, int *out, int N){
+void print_output(float *x, float *y, float *out, int N){
     for(int idx=0;idx<N;idx++){
-        printf("2 X %d + %d = %d\n", x[idx], y[idx], out[idx]);
+        printf("2 X %.1f + %.1f = %.1f\n", x[idx], y[idx], out[idx]);
     }
 }
 
 int main(){
     int N = NUM_BLOCKS * NUM_THREADS;
-    int size = N * sizeof(int);
-    int *x, *y, *out, *d_x, *d_y, *d_out;
+    int size = N * sizeof(float);
+    float *x, *y, *out, *d_x, *d_y, *d_out;
 
-    x = (int *)malloc(size); fill_array(x,N);
-    y = (int *)malloc(size); fill_array(y,N);
-    out = (int *)malloc(size);
+    x = (float *)malloc(size); fill_array(x,N);
+    y = (float *)malloc(size); fill_array(y,N);
+    out = (float *)malloc(size);
     
     cudaMalloc(&d_x, size); 
     cudaMalloc(&d_y, size);
